by robertragas: refactor social core dashboard homepage block
The dashboard core block is a menu link rendered on every page where the toolbar is present. This menulink is rendered dynamicly and because of the different menu context also hard to cache. This causes it to be loaded multiple times over multiple pages causing performance downgrade. This change refactors the homepage block to be set to a static link where we will redirect to the correct page when the link has been clicked. Elimating the need to execute this on every page.
