+++ modules/social_features/social_core/social_core.links.menu.yml
@@ -147,7 +147,9 @@ social_core.dashboard.appearance.change_colors_and_styling:
   route_name: social_core.dashboard.appearance.active_theme
   weight: -100
 
-social_core.dashboard.appearance.active_homepage_configuration:
+social_core.dashboard.appearance.active_homepage_configuration.link:
+  title: 'Customize home page'
   parent: social_core.dashboard.appearance
-  class: Drupal\social_core\Plugin\Menu\HomepageConfigurationHeroBlockLink
+  description: 'Change the image and text on the home page.'
+  url: 'internal:/admin/dashboard/homepage-settings'
   weight: -90
